Conflicts:
	app/src/main/AndroidManifest.xml
	app/src/main/java/com/example/dllo/myhunter/ui/activity/MainActivity.java
	app/src/main/java/com/example/dllo/myhunter/ui/adapter/RecommendedEditorAdapter.java
	app/src/main/java/com/example/dllo/myhunter/ui/adapter/RecommendedStoryAdapter.java
	app/src/main/java/com/example/dllo/myhunter/ui/fragment/FoundFragment.java
	app/src/main/java/com/example/dllo/myhunter/ui/fragment/RecommendedFragment.java
	app/src/main/res/layout/activity_asahunter.xml
	app/src/main/res/layout/activity_login.xml
	app/src/main/res/layout/fragment_found.xml
	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/dllo/myhunter/model/bean/RecommendedBean.java
@@ -111,7 +111,7 @@ public class RecommendedBean {
             }
 
             public static class HunListBean {
-                private double goodcomment_rate;
+                private int goodcomment_rate;
                 private String product_image;
                 private String name;
                 private String avatar_s;
@@ -130,11 +130,11 @@ public class RecommendedBean {
                     this.product_image = product_image;
                 }
 
-                public double getGoodcomment_rate() {
+                public int getGoodcomment_rate() {
                     return goodcomment_rate;
                 }
 
-                public void setGoodcomment_rate(double goodcomment_rate) {
+                public void setGoodcomment_rate(int goodcomment_rate) {
                     this.goodcomment_rate = goodcomment_rate;
                 }
